+++ src/pyramid_sqlalchemy/__init__.py
@@ -1,6 +1,5 @@
 import logging
 from sqlalchemy import engine_from_config
-from sqlalchemy.engine.interfaces import Dialect
 from pyramid_sqlalchemy.meta import metadata
 from pyramid_sqlalchemy.meta import BaseObject
 from pyramid_sqlalchemy.meta import Session
@@ -14,21 +13,19 @@ log = logging.getLogger('pyramid_sqlalchemy')
 def init_sqlalchemy(engine):
     """Initialise the SQLAlchemy models. This must be called before using
     using any of the SQLAlchemy managed the tables or classes in the model."""
-    if engine.dialect.__class__.do_begin_twophase is Dialect.do_begin_twophase:
-        Session.configure(twophase=False)
-        log.warn('SQL dialect %s does not support two-phase commits',
-                engine.dialect.name)
-    else:
-        log.debug('Enabling two-phase commit support.')
-        Session.configure(twophase=True)
     Session.configure(bind=engine)
     metadata.bind = engine
+
+
+def enable_sql_two_phase_commit(config, enable=True):
+        Session.configure(twophase=enable)
 
 
 def includeme(config):
     """'Convenience method to initialise all components of this
     :mod:`pyramid_sqlalchemy` package from a pyramid applicaiton.
     """
+    config.add_directive('enable_sql_two_phase_commit', enable_sql_two_phase_commit)
     engine = engine_from_config(config.registry.settings, 'sqlalchemy.')
     init_sqlalchemy(engine)