+++ docs/changes.rst
@@ -1,12 +1,13 @@
 Changelog
 =========
 
-1.1 - Unreleased
+1.1 - July 14, 2014
 -------------------
 
-- Add missing schema to Pyramid-URL. This broke ReST rendering on the PyPI page.
+- Add missing schema to the Pyramid-URL in the package description. This broke
+  ReST rendering on the PyPI page.
 
-- Add a new ``enable_sql_two_phase_commit()`` configuration directive enable
+- Add a new ``enable_sql_two_phase_commit()`` configuration directive to enable
   two-phase commit.
 
 - Enable foreign key constraint checking for SQLite in DatabaseTestCase.